cleaned up junk code a bit
