+++ gcodes.cpp
@@ -14,30 +14,8 @@
  * M4 int = laser pulse
  *
  */
-struct command{
-	char *code;
-	double x;
-	double y;
-};
-command setting(char *input){
-	command h;
-	h.code = input;
-	h.x = NULL;
-	h.y = NULL;
-}
-command instrument(char *input, double degree){
-	command i;
-	i.code = input;
-	i.x = degree;
-	i.y = NULL;
-}
 
-command movement(char *input, double x, double y){
-	command m;
-	m.code = input;
-	m.x = x;
-	m.y = y;
-}
+
 void Gcodes::GCodeParser(const char* taskname,
 			uint16_t stacksize, UBaseType_t priority,
 			QueueHandle_t UART_data){
@@ -60,20 +38,7 @@ void Gcodes::parse(){
 	double yCoord;
 	xQueueReceive(UART_dataQueue, &str, portMAX_DELAY);
 	command send;
-	if(str[0] == 'M'){
 
-	}
-	else if(str[0] == 'G'){
-		if(str[1] == '4'){
-			first = str.substr(3);
-			xCoord = std::string::stod(first, &sz );
-			send = instrument('M4', xCoord );
-
-		}
-
-	}
-	else{
-		//TODO error handling
-	}
+	/*************String parsing************/
 
 }
